Split childcare expense variables into separate files
