+++ openfisca_us/variables/expense/childcare/childcare_days_per_week.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class childcare_days_per_week(Variable):
+    value_type = float
+    entity = Person
+    label = u"Child care days per week"
+    definition_period = YEAR
