+++ openfisca_us/variables/expense/childcare/provider_type_group.py
@@ -1,5 +1,4 @@
-from openfisca_core.model_api import *
-from openfisca_us.entities import *
+from openfisca_us.model_api import *
 
 
 class ProviderTypeGroup(Enum):
@@ -18,26 +17,5 @@ class provider_type_group(Variable):
     # DCC_SACC is most common among provider types
     default_value = ProviderTypeGroup.DCC_SACC
     entity = Person
-    label = u"CCDF provider type group"
-    definition_period = YEAR
-
-
-class childcare_hours_per_week(Variable):
-    value_type = float
-    entity = Person
-    label = u"Child care hours per week"
-    definition_period = YEAR
-
-
-class childcare_hours_per_day(Variable):
-    value_type = float
-    entity = Person
-    label = u"Child care hours per day"
-    definition_period = YEAR
-
-
-class childcare_days_per_week(Variable):
-    value_type = float
-    entity = Person
-    label = u"Child care days per week"
+    label = u"Childcare provider type group"
     definition_period = YEAR
